+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed  
 
 - history.json, history2_air.json, and history2_sea.json files are now saved in a WaveLab folder under the UserProfile environment variable instead of the same directory as WaveLab.exe
+- Latitude and Longitude on graphs reduced to 4 decimal places
+- Instrument Error, Combined Instrument Error, Maximum Storm Tide Water Elevation, and Maximum Unfiltered Water Elevation reduced to 2 decimal places
 
 ### Deprecated